Fix some bugs based on server_testing branch.
The bugs below are all fixed.
=============================
1. What is the dbView-view data format? It seems as if it is like the
table below referring to Line
318 in HelloWorld.java. If so, why do you have to compare the item name
to “view” and use a “for” loop to set “viewString” in Line 337 and in
Line 441? Or is it permitted that the simpleDB could have two items
with the same item name, same attribute but with different value?
2. No matter there is a “dbView” in the simpleDB, the local view should
be set with its local IP, status and time. Why the subroutine
group.view() is called in both “if” part in Line 324 and “else” part in
Line 345? I think this subroutine should be used only once out of “if”
and “else” part. At the very beginning of getViewFromSimpleDB() is OK.
In fact I did it in GossipProtocolTest.java in this pattern.!
3. In Line 851, the timeStamp should be “long” type.!
4. In Line 348, after merging the view, why not update the simpleDB at
that time?!
5. In Line 790, I think if timeout, it is also necessary to run
“rpcSocket.close()” before “return null”.!
6. In Line 411, why remove server IP that has been chosen?